Fix syntax error due to use of non-transformable ES6 "new.target"
"new.target" is used to ensure that the base class cannot be instantiated directly, but this is an ES6 feature new to Node.js v5.0.0 that cannot be transformed directly by Babel v6 (https://github.com/babel/babel/issues/1088) and results in a syntax error in versions of Node.js <5.0.0. This change fixes the issue using this.constructor and ensures that the script properly runs on older versions of Node.js when Babel is used to build the dist files.

This also adjusts the Travis CI Node.JS versions to include Node.JS 4 and 0.12, which are listed as supported in the package.json file.

Fixes #45